+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [0.22.0-295b305](https://github.com/cluster-software/agent-crm-app/compare/v0.21.0-295b305...v0.22.0-295b305) (2026-05-28)
+
+
+### Features
+
+* import linkedin sync chunks ([54985f1](https://github.com/cluster-software/agent-crm-app/commit/54985f1372fa36a8bb0bc6f86823bbb4186f0531))
+
+## [0.21.0-295b305](https://github.com/cluster-software/agent-crm-app/compare/v0.20.3-295b305...v0.21.0-295b305) (2026-05-28)
+
+
+### Features
+
+* render structured email messages ([#133](https://github.com/cluster-software/agent-crm-app/issues/133)) ([e1a7f15](https://github.com/cluster-software/agent-crm-app/commit/e1a7f158ae18a62d0de3f402396a5054525d6a7d))
+
+## [0.20.3-295b305](https://github.com/cluster-software/agent-crm-app/compare/v0.20.2-295b305...v0.20.3-295b305) (2026-05-28)
+
+
+### Bug Fixes
+
+* keep table headers above avatars ([00b8f21](https://github.com/cluster-software/agent-crm-app/commit/00b8f210e5327de887142d189e8f9b90335fd6d3))
+* keep table headers above avatars ([e782117](https://github.com/cluster-software/agent-crm-app/commit/e782117b068bc31d664d38a3a93a8d38d537f15f))
+
 ## [0.20.2-295b305](https://github.com/cluster-software/agent-crm-app/compare/v0.20.1-295b305...v0.20.2-295b305) (2026-05-28)